framebuffer: update docs for _get_depth_texture()
This clarifies in the documentation for
cogl_framebuffer_get_depth_texture() that if a depth texture is added as
a layer to a pipeline and sampled then depth values should be read from
the red component and that all other components are undefined.

It depends on the version of OpenGL, and what profile is selected
whether sampling depth textures behaves like sampling a luminance
texture (L,L,L,1) or a component red texture (R,0,0,1).

In particular I was caught out by this issue in Rig where we were
referencing the .z component of shadow map textures, and this had worked
ok until updating to a recent version of mesa which has updated its
behaviour. The relevant mesa commit is: 678ac190a5a6fd39b8568587fac28

Reading the mailing list discussion that happened around that patch it
looks like Kenneth Graunke investigated the different variations so for
posterity it could be handy note them here:

Apparently it works like this:
- Core: RED
- Forward Compatible: RED
- Compatibility or <= 3.0: LUMINANCE
- ES2 with OES_depth_texture: LUMINANCE
- ES3 for unsized formats: RED
- ES3 for sized formats: LUMINANCE